DGN: Fix warnings on Windows
git-svn-id: https://svn.osgeo.org/gdal/trunk@32155 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/dgn/ogrdgnlayer.cpp
@@ -532,7 +532,7 @@ OGRFeature *OGRDGNLayer::ElementToFeature( DGNElemCore *psElement )
           DGNElemArc    *psArc = (DGNElemArc *) psElement;
           DGNPoint      asPoints[90];
           // TODO: std::abs abd std::max
-          int nPoints = MAX(1,ABS(psArc->sweepang) / 5) + 1;
+          int nPoints = static_cast<int>(MAX(1,ABS(psArc->sweepang) / 5) + 1);
           DGNStrokeArc( hDGN, psArc, nPoints, asPoints );
 
           poLine->setNumPoints( nPoints );
